requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
discord.py==1.2.4
idna==2.8
multidict==4.5.2
praw==6.4.0
prawcore==1.0.1
pycparser==2.19
PyNaCl==1.3.0
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.6
websocket-client==0.56.0
websockets==6.0
yarl==1.3.0
